News_xiaohongshu/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
httpx==0.27.2
pydantic==2.12.3
pydantic-settings==2.2.1
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
sqlalchemy==2.0.30
pymysql==1.1.1
alembic==1.13.1
apscheduler==3.10.4
python-dotenv==1.0.1
tavily-python==0.5.0
pytest==8.3.3
jinja2==3.1.4
python-multipart==0.0.9
